Renamed all active files to follow naming convention
