+++ React-Frontend/src/components/app.js
@@ -1,8 +1,7 @@
 import React, { Component } from 'react';
-import Sidebar from './sidebar';
 import ReactDOM from 'react-dom';
 
-var TAG = "MainApp | "
+var TAG = "MainApp | ";
 
 export default class App extends Component {
 
@@ -11,15 +10,6 @@ export default class App extends Component {
       <div>
         {this.props.children}
       </div>
-    )
+    );
   }
 }
-
-const bgColors = { "Default": "#81b71a",
-                    "Blue": "#00B1E1",
-                    "Cyan": "#37BC9B",
-                    "Green": "#8CC152",
-                    "Red": "#E9573F",
-                    "Yellow": "#F6BB42",
-                    "Orange": "#E29C6E"
-};
