--- React-Frontend/src/components/recommendations_list_item.js
@@ -1,28 +1,0 @@
-import React, { Component } from 'react';
-const TAG = 'RecommendationListItem | '
-
-export default class RecommendationsListItem extends Component {
-
-	renderRecommendation() {
-		console.log(TAG);
-		return (
-			<div className="recommendation-list-item">
-				<h3 className="results-subtitle">
-					{this.props.category}
-				</h3>
-				<img className="results-category-fitness-level" src={"../assets/level-"+this.props.fitness_level+".png"} />
-				<p className="results-recommendations-body">
-					{this.props.recommendation}
-				</p>
-			</div>
-		);
-	}
-
-	render() {
-		return (
-			<div>
-				{this.renderRecommendation()}
-			</div>
-		);
-	}
-}
